implement gcload and gctest in db_bench
gcload and gctest are for garbage collection testing;

gcload will fillseq an amount of key/values, then it will
delete a percentage of them, such as 20%.

gctest continues to fillrandom to the db created by gcload;
keys inserted in gctest will not overlapped with gcload, so that
gc will reclaim a fixed percentage of free space.

during gctest, gc of wisckey will be triggered, and the throughput
of gctest is the fillrandom throughput reported by db_bench.